Fix bug caused by mixing-in convenience functions
This bug was caused by upgrading lodash to the latest version.
Apprently, the mixin function wasn't behaving as intended in the
previous version, which we were using to our advantage.

We're now manually assigning the convenience functions to the object we
instantiate, and all our tests are once again passing.