Add ability to read UserData from a file
Add a new EC2 Instance convenience function for populating the UserData
section with the contents of a shell script passed to the function.

Currently, this returns a PROMISE so as not to block the event loop, but
it really funks up the fluent nature of Scenery. A TODO was added.

All unit tests have been checked in.