Make userDataFromFile fluent once more
This commit ensures reading files remains asynchronous so as not to block
the event loop, but makes userDataFromFile fluent by immeidatley
returning the Instance on which we're invoking the function.

Additionaly, the ability to pass a callback to the usereDataFromFile
function has been introduced to allow users to know when the task of
reading user data from a file has been complete. You can see how this is
used in the AWS_EC2_Instance_Convenience_Test.js unit test.